# executor: reuse chunk to reduce memory usage in union scan