<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/tidb/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
For a large transaction, union scan uses lots of memory which is harmful to the performance.

### What is changed and how it works?
Reuse chunk in when build added rows.

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test
 - Manual test (add detailed scripts or steps below)
execute 4900 update statements which affected 4900 rows  in one transaction.

| master | this PR |
| - | - |
| 17.944s | 7.079s |

Code changes

 - Has exported function/method change

Side effects

 - Increased code complexity

PTAL @coocood @lysu